DocSight/requirements.txt
streamlit
google-generativeai
langchain
PyPDF2
chromadb
langchain_google_genai
os